requirements.txt
numpy==1.19.2
streamlit==0.78.0
tensorflow==2.1.0
librosa==0.8.0
Pillow==8.1.2
